- Shift+Tab now switches sequencer and optional view - Bug fix with beat count < sig numerator in Loop Sampler
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@87 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/plugins/loop_sampler/LoopSampler.h
@@ -118,7 +118,7 @@ protected:
   float ReleaseCoef;
   float Envelope;
 
-  int	 BarCount;
+  double BarCount;
   int	 BeatCount;
   float	 Volume;
   float	 Attack;
